video_processor/requirements.txt
pytube==15.0.0
instaloader==4.10.2
python-dotenv==1.0.0
boto3==1.34.34
aiohttp==3.9.1
tenacity==8.2.3
python-magic==0.4.27